Refactor modules using rate limit handling
