+++ src/neo4j_graphrag/embeddings/vertexai.py
@@ -18,7 +18,7 @@ from typing import TYPE_CHECKING, Any, Optional
 
 from neo4j_graphrag.embeddings.base import Embedder
 from neo4j_graphrag.exceptions import EmbeddingsGenerationError
-from neo4j_graphrag.llm.rate_limit import RateLimitHandler
+from neo4j_graphrag.utils.rate_limit import RateLimitHandler
 
 try:
     from vertexai.language_models import TextEmbeddingInput, TextEmbeddingModel
