+++ src/neo4j_graphrag/llm/cohere_llm.py
@@ -20,7 +20,7 @@ from pydantic import ValidationError
 
 from neo4j_graphrag.exceptions import LLMGenerationError
 from neo4j_graphrag.llm.base import LLMInterface
-from neo4j_graphrag.llm.rate_limit import (
+from neo4j_graphrag.utils.rate_limit import (
     RateLimitHandler,
     rate_limit_handler,
     async_rate_limit_handler,
